+++ js/admin/src/components/BazaarPage.js
@@ -1,32 +1,27 @@
-import Component from 'flarum/Component';
-import ExtensionRepository from 'flagrow/bazaar/utils/ExtensionRepository';
-import ExtensionListItem from 'flagrow/bazaar/components/ExtensionListItem';
-import BazaarLoader from 'flagrow/bazaar/components/BazaarLoader';
+import Component from "flarum/Component";
+import ExtensionRepository from "flagrow/bazaar/utils/ExtensionRepository";
+import ExtensionListItem from "flagrow/bazaar/components/ExtensionListItem";
+import BazaarLoader from "flagrow/bazaar/components/BazaarLoader";
+import Button from "flarum/components/Button";
 
 export default class BazaarPage extends Component {
     init() {
         this.loading = m.prop(false);
         this.repository = m.prop(new ExtensionRepository(this.loading));
         this.repository().loadNextPage();
-        this.loader = BazaarLoader.component({loading: this.loading});
+        this.connected = app.data.settings['flagrow.bazaar.connected'] == 1 || false;
     }
 
     view() {
-        return (
-            <div className="ExtensionsPage Bazaar">
-                <div className="ExtensionsPage-header">
-                    <div className="container">
-                    </div>
-                </div>
-
-                <div className="ExtensionsPage-list">
-                    <div className="container">
-                        {this.items()}
-                    </div>
-                </div>
-                {this.loader}
-            </div>
-        );
+        return m('div', {className: 'ExtensionsPage Bazaar'}, [
+            m('div', {className: 'ExtensionsPage-header'}, [
+                m('div', {className: 'container'}, this.connectedHeader())
+            ]),
+            m('div', {className: 'ExtensionsPage-list'}, [
+                m('div', {className: 'container'}, this.items())
+            ]),
+            BazaarLoader.component({loading: this.loading})
+        ]);
     }
 
     items() {
@@ -35,5 +30,48 @@ export default class BazaarPage extends Component {
                 extension => ExtensionListItem.component({extension: extension, repository: this.repository})
             )
         ]);
+    }
+
+    connectedHeader() {
+        if (this.connected) {
+            return [
+                Button.component({
+                    className: 'Button Button--primary',
+                    icon: 'dashboard',
+                    children: app.translator.trans('flagrow-bazaar.admin.page.button.connected'),
+                    onclick: () => window.open('https://flagrow.io/home')
+                }),
+                m('p', [
+                    app.translator.trans('flagrow-bazaar.admin.page.button.connectedDescription')
+                ])
+            ]
+        }
+
+        return [
+            Button.component({
+                className: 'Button Button--primary',
+                icon: 'plug',
+                children: app.translator.trans('flagrow-bazaar.admin.page.button.connect'),
+                onclick: () => this.connect()
+            }),
+            m('p', [
+                app.translator.trans('flagrow-bazaar.admin.page.button.connectDescription')
+            ])
+        ]
+    }
+
+    connect() {
+        var popup = window.open();
+
+        app.request({
+            method: 'GET',
+            url: app.forum.attribute('apiUrl') + '/bazaar/connect'
+        }).then(response => {
+            if (response && response.redirect) {
+                popup.location = response.redirect;
+            } else {
+                popup.close()
+            }
+        });
     }
 }